Changed number value type from Integer to Long, to avoid issues with time values, which has been stored as Long already; while int time values would certainly last for a couple of millennia, the time values -inf and inf won't fit into an int

+++ src/pcm/state/MappedState.java
@@ -44,7 +44,7 @@ public class MappedState extends State {
     }
 
     @Override
-    public Integer get(Integer n) {
+    public Long get(Integer n) {
         if (mapping.containsKey(n)) {
             Items<Toys> items = mapping.get(n);
             boolean available = items.available().size() > 0;
@@ -78,8 +78,8 @@ public class MappedState extends State {
                 // 1:1 mapping
                 items.get(0).setAvailable(true);
             } else {
-                throw new IllegalStateException(n + "(" + items.toString()
-                        + ")" + ": Multiple-mapped values can only be unset");
+                throw new IllegalStateException(n + "(" + items.toString() + ")"
+                        + ": Multiple-mapped values can only be unset");
             }
         }
         super.set(n);
